add the stop-and make post-stop-frames (#76)
Closes #75 

this PR adds a piece of logic to ensure the the trigger is detected
correctly. the row for 'all the frames after the trigger' must be
separate from the row for 'detect the trigger'. While those two were
joined, at every `repeat` the trigger was checked, and that is not what
the desired behavior is.

Furthermore we need to check against `post_stop_frames > 1` so that we
do not pass `repeats=0`